% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{.draw_gumbel_uniforms}
\alias{.draw_gumbel_uniforms}
\title{Draw bivariate uniforms from a Gumbel--Hougaard copula}
\usage{
.draw_gumbel_uniforms(n, theta = 1)
}
\arguments{
\item{n}{integer; number of observations}

\item{theta}{numeric \eqn{>=} 1; Gumbel-Hougaard association parameter}
}
\value{
n x 2 matrix, each row (U1, U2), both in (0,1), with GH copula parameter = theta.
}
\description{
Draw bivariate uniforms from a Gumbel--Hougaard copula
}
\details{
This mirrors the construction of the \code{gumbel.jeong} function used in the \code{cWR} package. For even more details, see the manual for R package \code{gumbel}.
}
\keyword{internal}
