% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.f_D}
\alias{.f_D}
\title{Density of T_D}
\usage{
.f_D(y2, a, x, u, lambda_D, kappa_D, beta_A_D, beta_X_D, beta_U_D)
}
\arguments{
\item{y2}{numeric scalar; time for fatal event}

\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{lambda_D}{numeric scalar > 0; baseline hazard for fatal event}

\item{kappa_D}{numeric scalar > 0; Weibull shape for fatal event}

\item{beta_A_D}{numeric scalar; treatment coefficient for fatal event; can also be \code{NULL}}

\item{beta_X_D}{numeric vector; coefficients for x on fatal event; can also be \code{NULL}}

\item{beta_U_D}{numeric vector; coefficients for u on fatal event; can also be \code{NULL}}
}
\value{
numeric scalar; density value
}
\description{
Compute the density f_D(y2 | A,X,U) for the fatal event time T_D at y2.
}
\keyword{internal}
