% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{.generate_covariate_matrix}
\alias{.generate_covariate_matrix}
\title{Generate a covariate matrix with Normal and Bernoulli columns}
\usage{
.generate_covariate_matrix(
  N,
  p = NULL,
  dist = NULL,
  mu = NULL,
  sd = NULL,
  prob = NULL,
  prefix = "X"
)
}
\arguments{
\item{N}{integer; sample size.}

\item{p}{integer; number of covariates. If NULL or <= 0, returns NULL.}

\item{dist}{character vector length p; entries should be either "Normal" or "Bernoulli".
Designates the distributions each covariate belongs to.
If NULL and p > 0, defaults to all "Normal".}

\item{mu}{numeric vector length p; means for Normal covariates; indices should
match with the "Normal" indices in the \code{dist} argument; ignored for
"Bernoulli" indices. Can be NULL if there are no Normal covariates.}

\item{sd}{numeric vector length p; SDs for Normal covariates; indices should
match with the "Normal" indices in the \code{dist} argument; ignored for
"Bernoulli" indices. Can be NULL if there are no Normal covariates.}

\item{prob}{numeric vector length p; success probabilities for Bernoulli
covariates that must lie in (0,1); indices should match with the "Bernoulli"
indices in the \code{dist} argument; ignored for "Normal" indices. Can be
NULL if there are no Bernoulli covariates.}

\item{prefix}{character, prefix to be used for column names in the generated
dataset, i.e. "X" will generate columns "X_1", "X_2", ...}
}
\value{
A numeric matrix of dimension N x p, with columns named paste0(prefix, "_", 1:p).
Returns NULL if p is NULL or p == 0.
}
\description{
Generate a covariate matrix with Normal and Bernoulli columns
}
\keyword{internal}
