% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{.get_linear_predictor}
\alias{.get_linear_predictor}
\title{Extracts row-wise dot products}
\usage{
.get_linear_predictor(mat, coef, n)
}
\arguments{
\item{mat}{numeric matrix; represents covariates/features for each individual}

\item{coef}{numeric matrix; represents coefficients corresponding to each of the covariates in \code{mat}}

\item{n}{integer; sample size}
}
\value{
Returns a numeric vector of linear predictors, one per individual. Returns a vector of zeroes if there are no predictors.
}
\description{
Extracts row-wise dot products
}
\keyword{internal}
