% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_monte_carlo.R
\name{.get_WR}
\alias{.get_WR}
\title{Compute the "true" win ratio on a superpopulation.}
\usage{
.get_WR(treated, control)
}
\arguments{
\item{treated}{\code{data.frame}; containing the columns: \code{Y_D}, \code{delta_D}, \code{Y_H}, \code{delta_H}
for the treated arm.}

\item{control}{\verb{data frame}; same columns as \code{treated}}
}
\value{
A list with:
\itemize{
\item \code{treated_wins}, \code{control_wins}, \code{ties}
\item \code{total_pairs}
\item \code{pi_T}, \code{pi_C}
\item \code{WR} = \code{pi_T} / \code{pi_C}
}
}
\description{
After splitting the superpopulation into respective treatment and control \code{data.frame} objects,
computes the "population" win probabilities and the respective win ratio.
}
\details{
Uses the code infrastructure developed in the \code{WINS} package under the GPLv3 license.
Uses an all-against-all pairwise comparison approach against a fatal/non-fatal composite hierarchical outcome to determine winners/losers/and ties. Equal times are counted as ties.
}
