% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{.invphigumbel}
\alias{.invphigumbel}
\title{Gives the inverse generator function for the Gumbel-Hougaard Copula with parameter \code{theta}.}
\usage{
.invphigumbel(t, theta = 1)
}
\arguments{
\item{t}{numeric; dummy variable of the inverse generator \eqn{1/\phi}}

\item{theta}{numeric \eqn{>=} 1; Gumbel-Hougaard association parameter}
}
\value{
Gives the inverse generator.
}
\description{
Gives the inverse generator function for the Gumbel-Hougaard Copula with parameter \code{theta}.
}
\details{
Recreated from the \code{gumbel} package. See package documentation for details.
}
\keyword{internal}
