% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.Lambda_C}
\alias{.Lambda_C}
\title{Censoring hazard for random censoring}
\usage{
.Lambda_C(a, lambda_C, beta_A_C)
}
\arguments{
\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{lambda_C}{numeric scalar > 0; baseline censoring rate; can also be \code{NULL}}

\item{beta_A_C}{numeric scalar; treatment coefficient for censoring; can also be \code{NULL}}
}
\value{
numeric scalar; censoring hazard
}
\description{
Compute the hazard rate for random censoring. If \code{lambda_C} is NULL,
there is no random censoring and the hazard is zero.
}
\keyword{internal}
