% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.Lambda_H}
\alias{.Lambda_H}
\title{Weibull scale parameter for non-fatal event}
\usage{
.Lambda_H(a, x, u, lambda_H, beta_A_H, beta_X_H, beta_U_H)
}
\arguments{
\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{lambda_H}{numeric scalar > 0; baseline hazard rate}

\item{beta_A_H}{numeric scalar; treatment coefficient; can also be \code{NULL}}

\item{beta_X_H}{numeric vector; coefficients for x; can also be \code{NULL}}

\item{beta_U_H}{numeric vector; coefficients for u; can also be \code{NULL}}
}
\value{
numeric scalar; Weibull scale parameter for non-fatal event
}
\description{
Compute the scale parameter for the Weibull non-fatal event distribution.
}
\keyword{internal}
