% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.lambda_H_given_D}
\alias{.lambda_H_given_D}
\title{Conditional hazard lambda_{H|D}}
\usage{
.lambda_H_given_D(
  y1,
  y2,
  a,
  x,
  u,
  lambda_D,
  kappa_D,
  beta_A_D,
  beta_X_D,
  beta_U_D,
  lambda_H,
  kappa_H,
  beta_A_H,
  beta_X_H,
  beta_U_H,
  theta_copula,
  fd_step
)
}
\arguments{
\item{y1}{numeric scalar; time for non-fatal event}

\item{y2}{numeric scalar; time for fatal event}

\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{lambda_D}{numeric scalar > 0; baseline hazard for fatal event}

\item{kappa_D}{numeric scalar > 0; Weibull shape for fatal event}

\item{beta_A_D}{numeric scalar; treatment coefficient for fatal event; can also be \code{NULL}}

\item{beta_X_D}{numeric vector; coefficients for x on fatal event; can also be \code{NULL}}

\item{beta_U_D}{numeric vector; coefficients for u on fatal event; can also be \code{NULL}}

\item{lambda_H}{numeric scalar > 0; baseline hazard for non-fatal event}

\item{kappa_H}{numeric scalar > 0; Weibull shape for non-fatal event}

\item{beta_A_H}{numeric scalar; treatment coefficient for non-fatal event; can also be \code{NULL}}

\item{beta_X_H}{numeric vector; coefficients for x on non-fatal event; can also be \code{NULL}}

\item{beta_U_H}{numeric vector; coefficients for u on non-fatal event; can also be \code{NULL}}

\item{theta_copula}{numeric scalar >= 1; Gumbel–Hougaard parameter}

\item{fd_step}{numeric scalar > 0; finite difference step size}
}
\value{
numeric scalar; conditional hazard value
}
\description{
Compute the conditional hazard lambda_{H|D}(y1 | y2, A,X,U) via finite differences of the conditional survival G(y1, y2 | A,X,U).
This is basically a numerical method of calculating the numerator in the conditional hazard.
}
\keyword{internal}
