% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.linpred_D}
\alias{.linpred_D}
\title{Linear predictor for fatal event}
\usage{
.linpred_D(a, x, u, beta_A_D, beta_X_D, beta_U_D)
}
\arguments{
\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{beta_A_D}{numeric scalar; treatment coefficient; can also be \code{NULL}}

\item{beta_X_D}{numeric vector; coefficients for x; can also be \code{NULL}}

\item{beta_U_D}{numeric vector; coefficients for u; can also be \code{NULL}}
}
\value{
numeric scalar; linear predictor for the fatal event
}
\description{
Compute the linear predictor for the fatal event hazard given treatment, measured covariates x, and unmeasured covariates u.
}
\details{
One of x or u (and the corresponding beta parameter) can be \code{NULL}, but NOT both.
}
\keyword{internal}
