% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.linpred_H}
\alias{.linpred_H}
\title{Linear predictor for non-fatal event}
\usage{
.linpred_H(a, x, u, beta_A_H, beta_X_H, beta_U_H)
}
\arguments{
\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{beta_A_H}{numeric scalar; treatment coefficient; can also be \code{NULL}}

\item{beta_X_H}{numeric vector; coefficients for x; can also be \code{NULL}}

\item{beta_U_H}{numeric vector; coefficients for u; can also be \code{NULL}}
}
\value{
numeric scalar; linear predictor for the non-fatal event
}
\description{
Compute the linear predictor for the non-fatal event hazard given treatment, measured covariates x, and unmeasured covariates u.
}
\details{
One of x or u (and the corresponding beta parameter) can be \code{NULL}, but NOT both.
}
\keyword{internal}
