% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{.logistic_function}
\alias{.logistic_function}
\title{Logistic function as used in confounded treatment assignment}
\usage{
.logistic_function(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Element-wise logistic function.
}
\description{
Logistic function as used in confounded treatment assignment
}
\keyword{internal}
