% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.Q}
\alias{.Q}
\title{Censoring survival function Q(t | A=a)}
\usage{
.Q(t, a, lambda_C, beta_A_C)
}
\arguments{
\item{t}{numeric scalar; time}

\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{lambda_C}{numeric scalar > 0; baseline censoring rate; can also be \code{NULL}}

\item{beta_A_C}{numeric scalar; treatment coefficient for censoring; can also be \code{NULL}}
}
\value{
numeric scalar; censoring survival probability
}
\description{
Compute the censoring survival probability Q(t | A=a). If
\code{lambda_C} is NULL, there is no random censoring and Q(t|a) = 1.
}
\keyword{internal}
