% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.S_D}
\alias{.S_D}
\title{Marginal Weibull survival for fatal event}
\usage{
.S_D(t, a, x, u, lambda_D, kappa_D, beta_A_D, beta_X_D, beta_U_D)
}
\arguments{
\item{t}{numeric scalar; time}

\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{lambda_D}{numeric scalar > 0; baseline hazard rate}

\item{kappa_D}{numeric scalar > 0; Weibull shape parameter}

\item{beta_A_D}{numeric scalar; treatment coefficient; can also be \code{NULL}}

\item{beta_X_D}{numeric vector; coefficients for x; can also be \code{NULL}}

\item{beta_U_D}{numeric vector; coefficients for u; can also be \code{NULL}}
}
\value{
numeric scalar; survival probability
}
\description{
Compute the marginal survival function S_D(t | A,X,U) for the fatal event.
}
\keyword{internal}
