% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.S_DH}
\alias{.S_DH}
\title{Joint survival S_{D,H}(t_D, t_H | A,X,U) via Gumbel–Hougaard copula}
\usage{
.S_DH(
  y2,
  y1,
  a,
  x,
  u,
  lambda_D,
  kappa_D,
  beta_A_D,
  beta_X_D,
  beta_U_D,
  lambda_H,
  kappa_H,
  beta_A_H,
  beta_X_H,
  beta_U_H,
  theta_copula
)
}
\arguments{
\item{y2}{numeric scalar; time for fatal event T_D}

\item{y1}{numeric scalar; time for non-fatal event T_H}

\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{lambda_D}{numeric scalar > 0; baseline hazard for fatal event}

\item{kappa_D}{numeric scalar > 0; Weibull shape for fatal event}

\item{beta_A_D}{numeric scalar; treatment coefficient for fatal event; can also be \code{NULL}}

\item{beta_X_D}{numeric vector; coefficients for x on fatal event; can also be \code{NULL}}

\item{beta_U_D}{numeric vector; coefficients for u on fatal event; can also be \code{NULL}}

\item{lambda_H}{numeric scalar > 0; baseline hazard for non-fatal event}

\item{kappa_H}{numeric scalar > 0; Weibull shape for non-fatal event}

\item{beta_A_H}{numeric scalar; treatment coefficient for non-fatal event; can also be \code{NULL}}

\item{beta_X_H}{numeric vector; coefficients for x on non-fatal event; can also be \code{NULL}}

\item{beta_U_H}{numeric vector; coefficients for u on non-fatal event; can also be \code{NULL}}

\item{theta_copula}{numeric scalar >= 1; Gumbel–Hougaard parameter}
}
\value{
numeric scalar; joint survival probability
}
\description{
Compute the joint survival for (T_D, T_H) under a Gumbel–Hougaard copula.
When \code{theta_copula = 1}, this reduces to independence
S_D(t_D) * S_H(t_H).
}
\keyword{internal}
