% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_analytic.R
\name{.S_H}
\alias{.S_H}
\title{Marginal Weibull survival for non-fatal event}
\usage{
.S_H(t, a, x, u, lambda_H, kappa_H, beta_A_H, beta_X_H, beta_U_H)
}
\arguments{
\item{t}{numeric scalar; time}

\item{a}{numeric scalar, takes on value 1 or 0; treatment indicator}

\item{x}{numeric vector; measured covariates; can also be \code{NULL}}

\item{u}{numeric vector; unmeasured covariates; can also be \code{NULL}}

\item{lambda_H}{numeric scalar > 0; baseline hazard rate}

\item{kappa_H}{numeric scalar > 0; Weibull shape parameter}

\item{beta_A_H}{numeric scalar; treatment coefficient; can also be \code{NULL}}

\item{beta_X_H}{numeric vector; coefficients for x; can also be \code{NULL}}

\item{beta_U_H}{numeric vector; coefficients for u; can also be \code{NULL}}
}
\value{
numeric scalar; survival probability
}
\description{
Compute the marginal survival function S_H(t | A,X,U) for the non-fatal event.
}
\keyword{internal}
