% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Generate a single dataset under the specified semi-competing risk paradigm.}
\usage{
simulate_dataset(
  N,
  p = NULL,
  q = NULL,
  X_dist = NULL,
  U_dist = NULL,
  mu_X = NULL,
  sd_X = NULL,
  mu_U = NULL,
  sd_U = NULL,
  prob_X = NULL,
  prob_U = NULL,
  treat_assign = c("randomized", "confounded"),
  alpha_0 = 0,
  alpha_X = NULL,
  alpha_U = NULL,
  lambda_H,
  kappa_H,
  beta_A_H,
  beta_X_H = NULL,
  beta_U_H = NULL,
  lambda_D,
  kappa_D,
  beta_A_D,
  beta_X_D = NULL,
  beta_U_D = NULL,
  theta_copula = 1L,
  lambda_C = NULL,
  beta_A_C = NULL,
  phi_admin = NULL,
  incl_tfe = FALSE,
  incl_latent = FALSE
)
}
\arguments{
\item{N}{integer; sample size for this data set.}

\item{p}{integer; number of measured covariates.}

\item{q}{Integer; number of unmeasured covariates.}

\item{X_dist}{character vector length p; distribution type for each measured covariate, with entries being one of "Normal" or "Bernoulli". If NULL, all measured covariates are generated as Normal.}

\item{U_dist}{character vector length q; distribution type for each unmeasured covariate, with entries being one of "Normal" or "Bernoulli". If NULL, all unmeasured covariates are generated as Normal.}

\item{mu_X}{numeric vector length p; means for measured covariates.
Positions in the vector should match the positions in the \code{X_dist} argument whose entries are "Normal". All other entries will be ignored. Can be \code{NULL} if there are no "Normal" entries in \code{X_dist}.}

\item{sd_X}{numeric vector length p; SDs for measured covariates.
Positions in the vector should match the positions in the \code{X_dist} argument whose entries are "Normal". All other entries will be ignored. Can be \code{NULL} if there are no "Normal" entries in \code{X_dist}.}

\item{mu_U}{numeric vector length q; means for unmeasured covariates.
Positions in the vector should match the positions in the \code{U_dist} argument whose entries are "Normal". All other entries will be ignored. Can be \code{NULL} if there are no "Normal" entries in \code{U_dist}.}

\item{sd_U}{numeric vector length q; SDs for unmeasured covariates.
Positions in the vector should match the positions in the \code{U_dist} argument whose entries are "Normal". All other entries will be ignored. Can be \code{NULL} if there are no "Normal" entries in \code{U_dist}.}

\item{prob_X}{numeric vector length p; success probabilities for measured Bernoulli covariates. Must lie in (0,1).
Positions in the vector should match the positions in the \code{X_dist} argument whose entries are "Bernoulli". All other entries will be ignored. Can be \code{NULL} if there are no "Bernoulli" entries in \code{X_dist}.}

\item{prob_U}{numeric vector length q; success probabilities for unmeasured Bernoulli covariates. Must lie in (0,1).
Positions in the vector should match the positions in the \code{U_dist} argument whose entries are "Bernoulli". All other entries will be ignored. Can be \code{NULL} if there are no "Bernoulli" entries in \code{U_dist}.}

\item{treat_assign}{character; one of "randomized" or "confounded".}

\item{alpha_0}{numeric; intercept in treatment model if confounded.}

\item{alpha_X}{numeric vector length p; coefficients for X in treatment model.}

\item{alpha_U}{numeric vector length q; coefficients for U in treatment model.}

\item{lambda_H}{numeric; baseline hazard rate for time-to-non-fatal outcome.}

\item{kappa_H}{numeric; Weibull shape for time-to-non-fatal outcome. Becomes Exponentially distributed when equal to 1.}

\item{beta_A_H}{numeric; Treatment effect on time-to-non-fatal outcome.}

\item{beta_X_H}{numeric vector length p; Measured covariate effects on time-to-non-fatal outcome.}

\item{beta_U_H}{numeric vector length q; Unmeasured covariate effects on time-to-non-fatal outcome.}

\item{lambda_D}{numeric; baseline hazard rate for fatal outcome.}

\item{kappa_D}{numeric; Weibull shape parameter for marginal time-to-fatal outcome. Becomes Exponentially distributed when equal to 1.}

\item{beta_A_D}{numeric; treatment effect on fatal outcome; greater than zero decreases hazard.}

\item{beta_X_D}{numeric vector length p; measured covariate effects on time-to-fatal outcome.}

\item{beta_U_D}{numeric vector length q; unmeasured covariate effects on time-to-fatal outcome.}

\item{theta_copula}{numeric \eqn{>=} 1; Gumbel–Hougaard dependence parameter. No association between latent competing event times when equal to 1.}

\item{lambda_C}{numeric; baseline censoring rate.}

\item{beta_A_C}{numeric; treatment effect on censoring rate.}

\item{phi_admin}{numeric; administrative censoring time.}

\item{incl_tfe}{boolean: default FALSE; additionally includes calculated time-to-first-event variables in the output dataset.}

\item{incl_latent}{boolean; default FALSE; additionally includes latent event times in the output dataset.}
}
\value{
A \code{data.frame} object with columns:
\itemize{
\item \code{id}: unique subject identifier
\item \code{A}: treatment assignment
\item \code{Y_D}: observed time to fatal event
\item \code{delta_D}: oberved fatal event indicator
\item \code{Y_H}: observed time to non-fatal event
\item \code{delta_H}: observed non-fatal event indicator
\item \code{X_1}, ..., \code{X_p}: measured covariates
\item \code{U_1}, ..., \code{U_q}: unmeasured covariates
\item \code{Y_tfe}: observed time to first event (if \code{incl_tfe} is set to \code{TRUE})
\item \code{delta_tfe}: observed time to first event indicator (if \code{incl_tfe} is set to \code{TRUE})
\item\code{T_D}: latent time to fatal event (if \code{incl_latent} is set to \code{TRUE})
\item\code{T_H}: latent time to non-fatal event (if \code{incl_latent} is set to \code{TRUE})
}
}
\description{
Generate a single dataset of i.i.d. subjects under semi-competing risks, with:
\itemize{
\item Measured covariates \code{X} w/ dimension \code{p}.
\item Unmeasured covariates \code{U} w/ dimension \code{q}.
\item Treatment assignment \code{A}. Randomized or confounded.
\item Observed semi-competing outcomes: \code{Y_D}. \code{delta_D}, \code{Y_H}, \code{delta_H}.
\item Derived time-to-first-event outcomes \code{Y_tfe}, \code{delta_D}.
}
Under:
\itemize{
\item Latent time to fatal and non-fatal event: \code{T_D}, \code{T_H}.
\item Gumbel–Hougaard copula dependence with parameter \code{theta_copula}.
\item Random censoring time \code{C}.
\item Administrative censoring time \code{phi_admin}.
}
}
\details{
Weibull form:
S_D(t | A, X, U) = exp( - (Lambda_D * t)^kappa_D ),
where
Lambda_D = lambda_D * exp( - (beta_A_D * A +
beta_X_D^T X +
beta_U_D^T U) ).

Event time simulation via inverse transform:
T_D = (-log(V_D))^(1 / kappa_D) / Lambda_D,
given V_D ~ U(0, 1).

Joint dependence across (T_H, T_D) is induced by sampling
a bivariate (V_H, V_D) from a Gumbel–Hougaard copula with parameter
theta_copula, then applying the inverse Weibull transformations
separately for fatal and non-fatal events.
}
