% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_wr_monte_carlo.R
\name{true_wr_monte_carlo}
\alias{true_wr_monte_carlo}
\title{Monte Carlo approximation of the "true" causal win ratio as defined in Zhang 2021}
\usage{
true_wr_monte_carlo(N_super = 1e+06, seed = NULL, ...)
}
\arguments{
\item{N_super}{integer; size of the superpopulation (default \code{1e6})}

\item{seed}{integer or \code{NULL}; random seed}

\item{...}{Additional parameters passed on to \code{simulate_dataset()}.
See documentation of \code{simulate_dataset()} for a list of possible arguments.}
}
\value{
A list with:
\itemize{
\item \code{WR}: Monte Carlo estimate of the true win ratio
\item \code{pi_T}, \code{pi_C}: Population win probabilities
\item \code{treated_wins}, \code{control_wins}, \code{ties}, \code{total_pairs}: Respective counts during the pairwise comparison process
\item \code{N_super}: Superpopulation size used
}
}
\description{
Generates a large superpopulation under the specified data model,
and computes the empirical population-level win ratio using the observed outcomes.
}
\details{
This function computes the hierarchical WR on the observed times \code{Y_*} with indicators
\code{delta_*} of a massive simulated dataset. As \code{N_super} grows, this should converge to the analytic truth.
}
